fix(tasks): preserve legacy github token fallback
